requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
numpy
pandas
selenium==4.12.0
undetected-chromedriver==3.5.3
webdriver-manager==4.0.0
python-dotenv==1.0.0
requests==2.31.0
lxml
beautifulsoup4==4.12.2
flask-wtf==1.1.1
cryptography==41.0.3
APScheduler==3.10.4
flask
pandas
selenium
beautifulsoup4
requests
python-dotenv
PyPDF2==3.0.1
python-docx==0.8.11
scikit-learn>=1.5.0
spacy
nltk==3.8.1

google-generativeai

mysql-connector-python